fix: move enums to global file src/enums/index.ts
