+++ src/dataservices/item.service.ts
@@ -4,9 +4,9 @@ import type { FeedItem } from '../types/FeedItem.ts'
 import axios from '@nextcloud/axios'
 import _ from 'lodash'
 import { API_ROUTES } from '../types/ApiRoutes.ts'
+import { FEED_ORDER } from './../enums/index.ts'
 import store from './../store/app.ts'
 import feedstore from './../store/feed.ts'
-import { FEED_ORDER } from './feed.service.ts'
 
 export const ITEM_TYPES = {
 	FEED: 0,
