+++ src/enums/index.ts
@@ -1,0 +1,22 @@
+export enum DISPLAY_MODE {
+	DEFAULT = '0',
+	COMPACT = '1',
+	SCREENREADER = '2',
+}
+
+export enum FEED_ORDER {
+	DEFAULT = 0,
+	OLDEST = 1,
+	NEWEST = 2,
+}
+
+export enum FEED_UPDATE_MODE {
+	UNREAD = 0,
+	IGNORE = 1,
+}
+
+export enum SPLIT_MODE {
+	VERTICAL = '0',
+	HORIZONTAL = '1',
+	OFF = '2',
+}
